refactor(ui): TMS-2508 rename views
Renamed the views to make them a little more sensical.
